scripts/Functional_test/minimum_configuration_test/requirements.txt
crc==4.1.0
pyserial==3.5